pe_evaluate/requirements.txt
# AI模型评测系统依赖包

# 基础依赖
requests>=2.31.0
pyyaml>=6.0.1

# OpenAI SDK（DMXAPI 需要）
openai>=1.0.0

# 可选：更好的HTTP客户端
httpx>=0.25.0

# 可选：进度条显示
tqdm>=4.66.0

# 可选：更好的命令行输出
rich>=13.7.0

